Rename annotation
